pyproject.toml
[project]
name = "ocr-tech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2pdf>=0.1.8",
    "matplotlib>=3.10.5",
    "opencv-python>=4.12.0.88",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "python-docx>=1.2.0",
]
